asm_project/training/requirements.txt
# ASM Training Pipeline - Dependencies

## Core Dependencies (Required)
torch>=2.0.0
transformers>=4.35.0
peft>=0.6.0
accelerate>=0.24.0

## Document Processing
PyMuPDF>=1.23.0  # PDF processing
python-docx>=1.0.0  # DOCX processing
pandas>=2.0.0  # CSV processing
pytesseract>=0.3.10  # OCR for scanned documents
Pillow>=10.0.0  # Image processing

## Semantic Chunking
sentence-transformers>=2.2.0  # Multilingual embeddings
numpy>=1.24.0

## Optional: For quantized models (recommended for GPU training)
bitsandbytes>=0.41.0  # 8-bit/4-bit quantization
optimum>=1.14.0  # GPTQ/AWQ support

## Optional: For better training monitoring
wandb>=0.15.0  # Weights & Biases (optional)
tensorboard>=2.14.0  # TensorBoard (optional)
